foi realizado o refactory nos serializadores assim como foi feito a adequacao onde ele era utilizado
closes  #24
