+++ helpers/restfy.py
@@ -49,7 +49,7 @@ def make_rest(Serializer):
                     setattr(instance, key, value)
 
                 instance.save()
-                result = Serializer.serializer(instance)
+                result = Serializer.encode(instance)
         except Model.DoesNotExist:
             status = 404
             result = {
@@ -73,7 +73,7 @@ def make_rest(Serializer):
 
 
         try:
-            result = Serializer.serializer(
+            result = Serializer.encode(
                 Model.objects.get(id=id)
             )
         except Model.DoesNotExist:
@@ -143,7 +143,7 @@ def make_rest(Serializer):
                 response = HttpResponse(
                     content_type='application/json',
                     content=json.dumps([
-                        Serializer.serializer(state) for state in query
+                        Serializer.encode(state) for state in query
                     ])
                     )
             else:
@@ -165,12 +165,12 @@ def make_rest(Serializer):
         try:
             with transaction.atomic():
                 data = json.loads(request.body)
-                instance = Serializer.deserializer(data)
+                instance = Serializer.decode(data)
                 instance.save()
 
                 response = HttpResponse(
                     content=json.dumps(
-                        Serializer.serializer(instance)
+                        Serializer.encode(instance)
                     ),
                     status=201
                 )
